refactor(settings): extract SkillsHubSettings into directory-based module
- Convert single-file component (628 lines) to directory structure
- Extract types (SkillInfo, ExternalSource) to types.ts
- Extract all state and handlers to useSkillsHub hook
- Split UI into sub-components: ExternalSkillsSection, MySkillsSection,
  SkillCard, SearchInput, AddPathModal
- No behavior changes; all existing tests pass
